# docs/swagger.yaml
basePath: /
definitions:
  helpers.ErrorResponse:
    properties:
      error: {}
      info:
        type: string
      status:
        example: "false"
        type: string
    type: object
  model.HealthCheckSwaggerResponse:
    properties:
      info:
        type: string
      instance_id:
        type: string
      message:
        type: string
      service_name:
        type: string
      status:
        type: string
    type: object
  model.ShortenURLRequest:
    properties:
      exp:
        description: thời gian hết hạn (giây)
        example: 604800
        type: integer
      url:
        example: https://google.com
        type: string
    required:
    - url
    type: object
  model.ShortenURLSwaggerResponse:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  model.User:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  user.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  user.LoginSwaggerResponse:
    properties:
      data:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  user.RegisterUserRequest:
    properties:
      display_name:
        type: string
      email:
        type: string
      password:
        minLength: 8
        type: string
      username:
        type: string
    required:
    - display_name
    - email
    - password
    - username
    type: object
  user.RegisterUserSwaggerResponse:
    properties:
      data:
        $ref: '#/definitions/model.User'
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    REST API cho hệ thống GoBe K03 (health check, URL shortener, ...)
    Sử dụng Echo framework + Redis + Clean Architecture
  termsOfService: https://example.com/terms
  title: GoBe K03 Project API
  version: 1.0.3
paths:
  /v1/health-check:
    get:
      description: Kiểm tra trạng thái hoạt động của service và kết nối Redis
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.HealthCheckSwaggerResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
      summary: Check service health
      tags:
      - Health
  /v1/links/redirect/{code}:
    get:
      description: Redirect to original URL by short code
      parameters:
      - description: Short URL code
        in: path
        name: code
        required: true
        type: string
      responses:
        "302":
          description: Found - Redirect via Location header
          headers:
            Location:
              description: Original URL
              type: string
          schema:
            type: string
        "404":
          description: URL not found
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
      summary: Redirect short URL
      tags:
      - Links
  /v1/links/shorten:
    post:
      consumes:
      - application/json
      description: Generate a short code for a given URL
      parameters:
      - description: Shorten URL request (exp must be > 0)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ShortenURLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ShortenURLSwaggerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
      summary: Shorten URL
      tags:
      - Links
  /v1/users/login:
    post:
      consumes:
      - application/json
      description: Return JWT Token if Correct Incredentials
      parameters:
      - description: Input Required
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.LoginSwaggerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
      summary: Return JWT Token if Correct Incredentials
      tags:
      - User
  /v1/users/register:
    post:
      consumes:
      - application/json
      description: Create a new user with the provided information
      parameters:
      - description: User to create
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.RegisterUserSwaggerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
      summary: Create a new user
      tags:
      - User
schemes:
- http
swagger: "2.0"
tags:
- name: Health
- name: Links
- name: User